Separated the form to two component
