+++ src/app/app.module.ts
@@ -11,6 +11,7 @@ import {HttpClientModule} from "@angular/common/http";
 import { CreateEventComponent } from './recruiter/create-event/create-event.component';
 import { YourAppointmentsComponent } from './recruiter/your-appointments/your-appointments.component';
 import {RouterModule, Routes} from '@angular/router';
+import {CreateEventModule} from "./recruiter/create-event/create-event.module";
 
 const appRoute: Routes = [
   {path: 'appointments', component: YourAppointmentsComponent},
@@ -28,14 +29,15 @@ const appRoute: Routes = [
     CreateEventComponent,
     YourAppointmentsComponent
   ],
-  imports: [
-    BrowserModule,
-    NgbModule,
-    FormsModule,
-    HttpClientModule,
-    RouterModule.forRoot(appRoute),
-    ReactiveFormsModule
-  ],
+    imports: [
+        BrowserModule,
+        NgbModule,
+        FormsModule,
+        HttpClientModule,
+        RouterModule.forRoot(appRoute),
+        ReactiveFormsModule,
+        CreateEventModule
+    ],
   providers: [],
   bootstrap: [AppComponent]
 })
