+++ src/app/common/footer/footer.component.html
@@ -1,4 +1,4 @@
-<footer class="fixed-bottom">
+<footer class="">
   <div class="container-xxl h64 flex-center">
     <p>&copy; 2023 — EASYVISIT</p>
   </div>
