+++ src/app/recruiter/create-event/availability/availability.component.html
@@ -1,0 +1,77 @@
+<ng-template #content let-modal>
+  <div class="modal-header">
+    <h4 class="modal-title" id="modal-basic-title">Dodaj dostępność</h4>
+    <button type="button" class="btn-close" aria-label="Close" (click)="modal.dismiss('Cross click')"></button>
+  </div>
+  <div class="modal-body">
+    <form>
+      <div class="row">
+        <div class="col-6">
+          <div class="mb-3 form-outline">
+            <input type="time" id="startHour" name="startHour" formControlName="startHour"
+                   aria-describedby="" aria-labelledby=""
+                   required class="form-control">
+            <label for="startHour" class="form-label">Godzina od</label>
+            <p id="" class="form-text">Format hh:mm</p>
+          </div>
+        </div>
+        <div class="col-6">
+          <div class="mb-3 form-outline">
+            <input type="time" id="endHour" name="endHour" formControlName="endHour"
+                   aria-describedby="" aria-labelledby=""
+                   required class="form-control">
+            <label for="endHour" class="form-label">Godzina do</label>
+            <p id="endHourHelper" class="form-text">Format hh:mm</p>
+          </div>
+<!--          <div class="col-6">-->
+<!--            <div class="mb-3 form-outline">-->
+<!--              <input type="time" id="startHour" name="startHour" formControlName="startHour"-->
+<!--                     aria-describedby="" aria-labelledby=""-->
+<!--                     required class="form-control"-->
+<!--                     [class.is-invalid]="startHour.invalid && (startHour.dirty || startHour.touched)">-->
+<!--              <label for="startHour" class="form-label">Godzina od</label>-->
+<!--              <div *ngIf="startHour.invalid && (startHour.dirty || startHour.touched)" class="invalid-feedback">-->
+<!--                <p *ngIf="startHour.errors?.['required']" id="" class="mb-0">Podaj początkowy zakres</p>-->
+<!--              </div>-->
+<!--              <p id="" class="form-text">Format hh:mm</p>-->
+<!--            </div>-->
+<!--          </div>-->
+<!--          <div class="col-6">-->
+<!--            <div class="mb-3 form-outline">-->
+<!--              <input type="time" id="endHour" name="endHour" formControlName="endHour"-->
+<!--                     aria-describedby="" aria-labelledby=""-->
+<!--                     required class="form-control"-->
+<!--                     [class.is-invalid]="endHour.invalid && (endHour.dirty || endHour.touched)">-->
+<!--              <label for="endHour" class="form-label">Godzina do</label>-->
+<!--              <div *ngIf="endHour.invalid && (endHour.dirty || endHour.touched)" class="invalid-feedback">-->
+<!--                <p *ngIf="endHour.errors?.['required']" id="endHourError" class="mb-0">Podaj końcowy zakres</p>-->
+<!--              </div>-->
+<!--              <p id="endHourHelper" class="form-text">Format hh:mm</p>-->
+<!--            </div>-->
+        </div>
+      </div>
+    </form>
+  </div>
+  <div class="modal-footer">
+    <button type="button" class="btn btn-outline-dark" (click)="modal.close('Save click')">Save</button>
+  </div>
+</ng-template>
+<!--Koniec modala-->
+
+<p class="h5 mb-4">Dodaj swoją dostępność dla tego “eventu”</p>
+<section class="row calendar">
+  <section class="col-12 col-sm-6 col-lg-4" *ngFor="let item of items">
+    <div class="c-card">
+      <div class="c-card-header">
+        <div><p class="mb-0 ps-2">{{ item.name }}</p></div>
+        <button type="button" class="btn btn-outline-primary" title="Dodaj dostępność" (click)="open(content)">Dodaj</button>
+      </div>
+      <div class="c-card-body">
+        <div class="slot" *ngFor="let hour of item.hour">
+          <div><p class="mb-0 ps-2">{{hour}}</p></div>
+          <button type="button" class="btn btn-outline-primary">Usuń</button>
+        </div>
+      </div>
+    </div>
+  </section>
+</section>
