+++ src/app/recruiter/create-event/availability/availability.component.scss
@@ -1,0 +1,33 @@
+.calendar {
+  gap:32px 0;
+  .c-card {
+    overflow: hidden;
+    border-radius: 0.25rem;
+    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
+    .c-card-header {
+      display: flex;
+      flex-flow: row;
+      justify-content: space-between;
+      align-items: center;
+      background-color: white;
+      padding: 0.5rem;
+      font-weight: 500;
+      font-size: 1rem;
+      line-height: 1.25;
+      .smaller {
+        font-size: 0.85rem;
+      }
+    }
+    .c-card-body {
+      .slot {
+        background-color: #D9D9D9;
+        padding: 0.5rem;
+        display: flex;
+        flex-flow: row;
+        justify-content: space-between;
+        align-items: center;
+        border-bottom: 1px solid #B5B0C5;
+      }
+    }
+  }
+}
