+++ src/app/recruiter/create-event/create-event.module.ts
@@ -1,12 +1,23 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { DefiningEventComponent } from './defining-event/defining-event.component';
+import { AvailabilityComponent } from './availability/availability.component';
+import {ReactiveFormsModule} from "@angular/forms";
 
 
 
 @NgModule({
-  declarations: [],
+    declarations: [
+        DefiningEventComponent,
+        AvailabilityComponent
+    ],
+  exports: [
+    AvailabilityComponent,
+    DefiningEventComponent
+  ],
   imports: [
-    CommonModule
+    CommonModule,
+    ReactiveFormsModule
   ]
 })
 export class CreateEventModule { }
